+++ plugins/recruit-resources/src/components/AvatarEditor.svelte
